+++ packages/cli/package.json
@@ -67,6 +67,7 @@
   "devDependencies": {
     "@react-native-community/cli-types": "^11.0.0",
     "@rnx-kit/scripts": "*",
+    "@types/connect": "^3.4.36",
     "@types/node-fetch": "^2.6.5",
     "@types/qrcode": "^1.4.2",
     "eslint": "^8.0.0",
@@ -81,6 +82,7 @@
   },
   "depcheck": {
     "ignoreMatches": [
+      "connect",
       "readline"
     ]
   },
